requirements.txt
fastapi 
uvicorn 
sqlmodel 
sqlalchemy 
pydantic 
python-dotenv
pyodbc
fastapi[all] 
python-jose[cryptography] 
passlib[bcrypt]